"finished" additional task 1
I did this buy going and making a small edit to the vehicle class in TicketLibrary. Then I rebuilt the DLL and reimplemented it in the other libraries.
As a result, A customer in the customer library, can now get the price of all trips by calling an object from TripsList. Let car1 be one of the trips in the list, car1.FinalPrice will return the final price of the trip, so that would be with or without discount, and whatever type of trip or vehicle it was.